incrementing version and release notes to 3.2.0.
Changing printing to use wxFileStream, and ostream instead of FILE*.  
Moving some macros (like PNT_LABEL) to be an internal const.
Getting rid of error string.  No longer necessary, really.
Config should be using doubles instead of floats, and it should use them in a different way.  This was causing some doubles to be interpreted as longs.

git-svn-id: http://svn.code.sf.net/p/calchart/svn/trunk@268 eb3ee358-668b-4427-8ffb-f370b80720d0