I created a readme file that contains the steps I followed to get calchart to compile on my ubuntu system. If compiling on another debian/linux/unix system, take the instructions with a grain of salt.
I also modified the makefile to work on my system; the changes hopefully shouldn't break functionality for others. I modified the file to ignore the 255 exit code being outputted by tex2rtf on my system and to generate the html help files as part of make all, as they are essential for the program.


git-svn-id: http://svn.code.sf.net/p/calchart/svn/trunk@297 eb3ee358-668b-4427-8ffb-f370b80720d0