This adds all sorts of music-related features to calchart: 1. The ability to play music behind a show. 2. The ability to specify information about the music behind a show (which will be useful in automatic continuity file generation). The information includes song names, beats per bar, and tempo shifts. 3. The ability to specify where the beats are in a music file, like an mp3, and export a json file, in the format designed for the CalChart Viewer, that will record where all the beats are. 4. A way to import a json beats file and a music file to play behind a show.
This is all up and running, but some of it is a little messy, and a lot of
it is pretty buggy (i didn't do so much checking for bad inputs). So I'll
need to go back and revise that stuff before I submit a pull request, but
here's the first version, just so you all can see what I'm up to.