Wide variety of changes to make modernize CalChart.
Working towards a more unified syntax for easier reading.

Consolidating many of the ancillary data objects for the Show into the Doc to allow greater understanding on what is part of the larger show instead of transient state.
Making FieldCanvas easier to understand by changing wording and objects to be consistent (ie, lasso is a type of shape, and how to check if something is in a polygon is a freestanding object).

using auto in more places as a more modern programming technique.

Changing naming convention of files to CamelCase style that is consistent with the name of the class they are containing inside.