Issue #451: Have the PrintContinuity Draw commands be generated by CalChart Core.
Adding Stack concept to drawing.  This allows a concept of a HStack/VStack for Horizontal and Vertical layouts.
Adding a more flexible way to do layout in terms of stack and surfaces.
Supports tabs with variable stops.
Adding a Playground for testing stack based drawing.
Incorporating Ranges and Coroutines to simplify code.
Adding in a new way to parse using coroutines.
Using the CalChart Point to draw the continuity.