Issue #444: Continuity view on main frame looks wrong (#445)
This was caused by a problem with wxWidgets.  Updating to a branch of 3.2.4 with the fix.

Also using this as a moment to update the way we do drawing.  Using more of the CalChart::DrawCommand concept to be more portable.

Using ranges whenever possible.
De-fanging operator+ for DrawCommands by making it explicitly create a range.

Updating xcode for ranges.