Issue #608: Add the ability to draw a curve in CalChart (#609)
Adding curves to CalChart that are rendered using Catmull-Rom algorithms.
Currently backed behind a feature flag.
This just adds the ability to draw.  They aren't saved as part of the show, nor do marchers interact with them.
There is a draw icon and a way to edit their colors in the configuration preference editor.
Adding default color for curves when we draw.
Added unit tests.

You can escape/delete to stop drawing curve.