Issue #608: Add the ability to draw a curve in CalChart
Adding ability to assign points to curves
Points get inserted to a curve and then you can move the curve.
Unit tests restructured to have better coverage of sheets.
It's so beautiful.

Fixing up some of the details of putting things on curves.

Starting on the way to get a dialog to edit the show points.

Working on a curve editor.